fix: resolve web lint warnings and desktop dead code warning
